fix(tests,libs,tools): resolve the bugprone and analyzer findings
Real defects the widened scope found:
- tests/fixtures/epub/epub_probe.c leaked the FILE* and the whole image
  buffer on the short-read path and never checked fseek/ftell/fread/malloc.
  Rewritten around a probe_slurp() helper that releases everything on every
  path.
- tests/host/exfat_fs_test.c had the same unchecked-stream pattern and built
  its ra8_fs_backend_t positionally; now checked and designated.
- The three KAT hex decoders moved from sscanf("%2x") to strtoul with an
  end-pointer check -- sscanf cannot report a conversion error, so a
  malformed vector silently decoded as 0x00.
- Array-bound gaps closed with real preconditions: cov_append() carries the
  destination capacity (66 call sites), tap() bounds the key index, and the
  RTT ring mirror rejects a write index past its storage.
- ra8_sdhi_bus_width_t gains k_ra8_sdhi_bus_width_unset = 0 (the sdcard
  config already documented 0 as "leave the bus alone" but had no
  enumerator), ra8_mipi_dsi_dt_t gains the reserved 0x00 data type, and
  test_ra8_ptp.c names its clock class in the initialiser.
- Struct comparisons that went through memcmp now compare members:
  ra8_touch_cal_matrix_t is all float, ra8_tileatlas_info_t carries padding.
- Duplicate switch arms merged, nested conditional operators unnested (two
  extracted into named helpers), casts through void* replaced with direct
  casts, and a redundant  arm dropped.
- tests/support/*.h helpers become static inline: a helper that a particular
  TU in the family does not call is not dead code, and static inline is the
  idiomatic header form.

Waivers, each with the reason written above it: the rand-stub tests (the
unit under test IS rand()/srand(), and a fixed seed is what makes the
sequence assertable), the DI mocks and the headless board_view backend
whose signatures are fixed by the typedef/header they implement, the const
ra8_psa_key_t handles (the typedef is itself a pointer), and
__executable_start (spelled by the linker).

Refs #296

+++ tools/board_sim/src/sim_seam_mve.c
@@ -218,7 +218,7 @@ static void on_mve_vstrw(uc_engine* uc, uint64_t address, uint32_t size, void* u
   uint8_t buf[k_mve_q_bytes];
   (void)memcpy(buf, &lo, sizeof(lo));
   (void)memcpy(buf + sizeof(lo), &hi, sizeof(hi));
-  (void)uc_mem_write(uc, (uint64_t)(base + off), buf, (size_t)k_mve_q_bytes);
+  (void)uc_mem_write(uc, (uint64_t)base + (uint64_t)off, buf, (size_t)k_mve_q_bytes);
   const uint32_t next = (uint32_t)address + (uint32_t)k_mve_insn_len;
   (void)uc_reg_write(uc, UC_ARM_REG_PC, &next);
   s_mve_emulated++;
